refactor(host): split app service workspace git module
